refactor: Line에 포함된 Section을 연결하는 역할을 SectionDao -> LineManager로 이동한다
