+++ src/main/java/subway/domain/Section.java
@@ -43,7 +43,7 @@ public class Section {
         requestSection.upSection = this;
     }
 
-    public void connectSection(Section requestSection) {
+    void connectSection(Section requestSection) {
         Assert.notNull(requestSection, () -> "requestSection은 null이 될 수 없습니다");
         SectionConnector.findSectionConnector(this, requestSection)
                 .ifPresentOrElse(sectionConnector1 -> sectionConnector1.connectSection(this, requestSection),
